Make Home into folder for better redux usage #5
